feat: FF 7 efeitos, endstop virtual, calibração persistente e export pickle
- G923: expande de 4 para 8 efeitos FF simultâneos (spring, damper, friction,
  inertia, constant, rumble, periodic, endstop)
- Adiciona FF_RUMBLE (impactos), FF_PERIODIC (RPM motor), FF_INERTIA (peso)
- Endstop virtual: trava progressiva nos limites calibrados da direção
- Calibração: desativa endstop automaticamente durante calibração de steering
- Calibração: reaplicada após find_device() no startup (fix sobrescrita)
- Mínimos sempre ativos: spring 5%, friction 3%, inertia 5%, periodic idle
- Max Force: threshold 25% com indicador vermelho na UI
- Export pickle: adiciona campos FF dinâmicos e inputs G923
- Motor RPi: arredonda PWM para inteiro (floor)

+++ raspberry/motor_manager.py
@@ -500,8 +500,8 @@ class MotorManager:
         # Obter limitador da marcha atual
         max_pwm = gear_limiters.get(self.current_gear, 40)
 
-        # Mapeia throttle (0-100%) para (0-limitador%)
-        final_pwm = (throttle_percent / 100.0) * max_pwm
+        # Mapeia throttle (0-100%) para (0-limitador%), arredonda para baixo
+        final_pwm = int((throttle_percent / 100.0) * max_pwm)
 
         return final_pwm
